refactor(core): collapse dual scrapbook resolvers; split scrapbook.ts (#192, #202)
`scrapbookDirForEntry(...)` is now the only public scrapbook-dir
resolver. The slug-template family (`scrapbookDir`, `scrapbookFilePath`,
`createScrapbookMarkdown`, `saveScrapbookFile`, `renameScrapbookFile`,
`deleteScrapbookFile`, `writeScrapbookUpload`) is no longer exported
from `@deskwork/core/scrapbook` — those names move to private helpers
(`_scrapbookDirSlug`, etc.) inside the module family. When
`scrapbookDirForEntry` is called with an entry whose `id` is missing
(legacy / pre-sidecar callers), it falls back internally to the
private slug-template path; external callers can no longer reach the
slug-only entry point at all.

The asymmetric resolver pair was the bug seam #191 fixed at the route
layer; #192 closes the seam at the API surface so it can't reappear.

**Folded in #202: split `packages/core/src/scrapbook.ts`**

909 lines → 94-line barrel + 9 sibling modules under
`packages/core/src/scrapbook/`:

- `types.ts` (54): public types — ScrapbookItem, ScrapbookSummary,
  ScrapbookLocation, etc.
- `validation.ts` (111): assertSlug, assertFilename, slugSegments,
  isNestedSlug, classify.
- `paths.ts` (203): public scrapbookDirAtPath /
  scrapbookDirForEntry / scrapbookFilePathAtDir; private
  _scrapbookDirSlug / _scrapbookFilePathSlug fallback.
- `listing.ts` (202): listScrapbook / listScrapbookAtDir /
  listScrapbookForEntry / countScrapbook / countScrapbookForEntry.
- `read.ts` (105): readScrapbookFile / readScrapbookFileAtDir /
  readScrapbookFileForEntry.
- `crud-at-dir.ts` (148): public entry-aware mutation primitives
  (createScrapbookMarkdownAtDir, saveScrapbookFileAtDir, etc. — added
  in #191).
- `crud-slug.ts` (149): INTERNAL slug-template CRUD with underscore
  prefixes; only used by seed.ts.
- `seed.ts` (61): seedScrapbookReadme (slug-keyed scaffolding helper).
- `format.ts` (33): formatRelativeTime, formatSize.

Each new module is well under the 300-line cap. The barrel
`scrapbook.ts` is pure re-exports of the public surface only.

**Caller migrations** (every grep hit in production code):

- `packages/studio/src/routes/scrapbook-mutation-envelope.ts` —
  dropped `scrapbookDir` import; slug-mode `resolveScrapbookDir` now
  routes through `scrapbookDirForEntry({ slug })` (the resolver's
  internal fallback covers the legacy path).
- `packages/studio/src/routes/scrapbook-mutation-dispatch.ts` —
  dropped imports of slug-template CRUD helpers + `scrapbookFilePath`.
  Dispatchers collapsed to the `*AtDir` path; the slug branch is
  gone (the dir resolver does the work upstream). 231 → 121 lines.
- `packages/studio/src/pages/scrapbook.ts` — replaced two
  `scrapbookFilePath(...)` calls with `scrapbookDirAtPath` +
  `scrapbookFilePathAtDir`.
- `packages/studio/src/routes/scrapbook-mutations.ts` /
  `scrapbook-file.ts` — header docstrings updated.

No CLI migrations needed; `@deskwork/cli` had zero direct callers of
the slug-template helpers.

**Tests rewritten:**

- `packages/core/test/scrapbook.test.ts` — exercises the new public
  surface. Slug-only fallback coverage preserved by exercising
  `scrapbookDirForEntry({ slug })`.

Workspace tests:
- @deskwork/core: 480 passed (unchanged)
- @deskwork/cli: 191 passed, 29 skipped (unchanged)
- @deskwork/studio: 361 passed, 11 skipped (unchanged)

**Live verification:**

Restarted the studio against the workspace dist (post-refactor).
POST'd `/api/dev/scrapbook/create` with `entryId` envelope against the
PRD entry; file landed at the entry-aware path. Studio review surface
still renders the PRD correctly. Test artifact removed.

Workplan + README status table updated.

Out-of-scope follow-up (intentionally NOT touched):
- `plugins/deskwork-studio/public/src/scrapbook-mutations.ts` (600
  lines, also flagged in #202) — separate refactor, doesn't share
  the export-surface concern that motivated the core split.

Closes (pending v0.16.0 marketplace-walk verification): #192, #202.
