+++ packages/core/src/scrapbook/format.ts
@@ -1,0 +1,33 @@
+/**
+ * Formatting helpers for the studio chip / viewer.
+ *
+ * Pure stringly-typed conversions — no fs, no config — so the studio
+ * client can import these without pulling in the rest of the scrapbook
+ * module family.
+ */
+
+export function formatRelativeTime(iso: string, now: Date = new Date()): string {
+  const then = new Date(iso).getTime();
+  const diff = now.getTime() - then;
+  if (diff < 0) return 'just now';
+  const s = Math.floor(diff / 1000);
+  if (s < 60) return `${s}s ago`;
+  const m = Math.floor(s / 60);
+  if (m < 60) return `${m}m ago`;
+  const h = Math.floor(m / 60);
+  if (h < 48) return `${h}h ago`;
+  const d = Math.floor(h / 24);
+  if (d < 14) return `${d}d ago`;
+  const w = Math.floor(d / 7);
+  if (w < 9) return `${w}w ago`;
+  const months = Math.floor(d / 30);
+  if (months < 18) return `${months}mo ago`;
+  const y = Math.floor(d / 365);
+  return `${y}y ago`;
+}
+
+export function formatSize(bytes: number): string {
+  if (bytes < 1024) return `${bytes} B`;
+  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
+  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+}
