+++ packages/studio/src/routes/scrapbook-mutation-dispatch.ts
@@ -1,10 +1,13 @@
 /**
  * Per-mode dispatch helpers for the studio scrapbook mutation routes.
  *
- * Each helper takes the parsed envelope + the mutation arguments and
- * dispatches to either the slug-template helper (slug mode) or the
- * `*AtDir` helper after resolving the entry-aware dir (entry mode). The
- * route handler stays thin — read JSON, parse envelope, validate route-
+ * Post-#192: both modes (`entry` and `slug`) route through the entry-
+ * aware `*AtDir` family after `resolveScrapbookDir` produces an absolute
+ * scrapbook directory. The slug-mode resolution still goes through
+ * `scrapbookDirForEntry` — passing `{ slug }` without an id triggers
+ * the internal slug-template fallback inside the resolver — but the
+ * public CRUD surface is unified on the `*AtDir` helpers. The route
+ * handler stays thin: read JSON, parse envelope, validate route-
  * specific args, dispatch.
  *
  * Extracted from `scrapbook-mutations.ts` (#191) to keep the route file
@@ -12,17 +15,11 @@
  */
 
 import {
-  createScrapbookMarkdown,
   createScrapbookMarkdownAtDir,
-  deleteScrapbookFile,
   deleteScrapbookFileAtDir,
-  renameScrapbookFile,
   renameScrapbookFileAtDir,
-  saveScrapbookFile,
   saveScrapbookFileAtDir,
-  scrapbookFilePath,
   scrapbookFilePathAtDir,
-  writeScrapbookUpload,
   writeScrapbookUploadAtDir,
   type ScrapbookItem,
 } from '@deskwork/core/scrapbook';
@@ -39,66 +36,24 @@ export async function saveDispatch(
   filename: string,
   bodyText: string,
 ): Promise<ScrapbookItem> {
-  if (env.mode === 'entry') {
-    const dir = await resolveScrapbookDir(ctx, env);
-    const abs = scrapbookFilePathAtDir(dir, filename, { secret: env.secret });
-    if (!existsSync(abs)) {
-      if (filename.endsWith('.md')) {
-        return createScrapbookMarkdownAtDir(dir, filename, bodyText, {
-          secret: env.secret,
-        });
-      }
-      return writeScrapbookUploadAtDir(
-        dir,
-        filename,
-        Buffer.from(bodyText, 'utf-8'),
-        { secret: env.secret },
-      );
-    }
-    return saveScrapbookFileAtDir(dir, filename, bodyText, {
-      secret: env.secret,
-    });
-  }
-  // Slug-template fallback path.
-  const abs = scrapbookFilePath(
-    ctx.projectRoot,
-    ctx.config,
-    env.site,
-    env.slug,
-    filename,
-    { secret: env.secret },
-  );
+  const dir = await resolveScrapbookDir(ctx, env);
+  const abs = scrapbookFilePathAtDir(dir, filename, { secret: env.secret });
   if (!existsSync(abs)) {
     if (filename.endsWith('.md')) {
-      return createScrapbookMarkdown(
-        ctx.projectRoot,
-        ctx.config,
-        env.site,
-        env.slug,
-        filename,
-        bodyText,
-        { secret: env.secret },
-      );
+      return createScrapbookMarkdownAtDir(dir, filename, bodyText, {
+        secret: env.secret,
+      });
     }
-    return writeScrapbookUpload(
-      ctx.projectRoot,
-      ctx.config,
-      env.site,
-      env.slug,
+    return writeScrapbookUploadAtDir(
+      dir,
       filename,
       Buffer.from(bodyText, 'utf-8'),
       { secret: env.secret },
     );
   }
-  return saveScrapbookFile(
-    ctx.projectRoot,
-    ctx.config,
-    env.site,
-    env.slug,
-    filename,
-    bodyText,
-    { secret: env.secret },
-  );
+  return saveScrapbookFileAtDir(dir, filename, bodyText, {
+    secret: env.secret,
+  });
 }
 
 export async function renameInPlace(
@@ -107,21 +62,10 @@ export async function renameInPlace(
   oldName: string,
   newName: string,
 ): Promise<ScrapbookItem> {
-  if (env.mode === 'entry') {
-    const dir = await resolveScrapbookDir(ctx, env);
-    return renameScrapbookFileAtDir(dir, oldName, newName, {
-      secret: env.secret,
-    });
-  }
-  return renameScrapbookFile(
-    ctx.projectRoot,
-    ctx.config,
-    env.site,
-    env.slug,
-    oldName,
-    newName,
-    { secret: env.secret },
-  );
+  const dir = await resolveScrapbookDir(ctx, env);
+  return renameScrapbookFileAtDir(dir, oldName, newName, {
+    secret: env.secret,
+  });
 }
 
 export interface CrossSectionPaths {
@@ -136,30 +80,10 @@ export async function resolveCrossSectionPaths(
   newName: string,
   toSecret: boolean,
 ): Promise<CrossSectionPaths> {
-  if (env.mode === 'entry') {
-    const dir = await resolveScrapbookDir(ctx, env);
-    return {
-      srcAbs: scrapbookFilePathAtDir(dir, oldName, { secret: env.secret }),
-      dstAbs: scrapbookFilePathAtDir(dir, newName, { secret: toSecret }),
-    };
-  }
+  const dir = await resolveScrapbookDir(ctx, env);
   return {
-    srcAbs: scrapbookFilePath(
-      ctx.projectRoot,
-      ctx.config,
-      env.site,
-      env.slug,
-      oldName,
-      { secret: env.secret },
-    ),
-    dstAbs: scrapbookFilePath(
-      ctx.projectRoot,
-      ctx.config,
-      env.site,
-      env.slug,
-      newName,
-      { secret: toSecret },
-    ),
+    srcAbs: scrapbookFilePathAtDir(dir, oldName, { secret: env.secret }),
+    dstAbs: scrapbookFilePathAtDir(dir, newName, { secret: toSecret }),
   };
 }
 
@@ -168,19 +92,8 @@ export async function deleteDispatch(
   env: ParsedEnvelope,
   filename: string,
 ): Promise<void> {
-  if (env.mode === 'entry') {
-    const dir = await resolveScrapbookDir(ctx, env);
-    deleteScrapbookFileAtDir(dir, filename, { secret: env.secret });
-    return;
-  }
-  deleteScrapbookFile(
-    ctx.projectRoot,
-    ctx.config,
-    env.site,
-    env.slug,
-    filename,
-    { secret: env.secret },
-  );
+  const dir = await resolveScrapbookDir(ctx, env);
+  deleteScrapbookFileAtDir(dir, filename, { secret: env.secret });
 }
 
 export async function createDispatch(
@@ -189,21 +102,10 @@ export async function createDispatch(
   filename: string,
   bodyText: string,
 ): Promise<ScrapbookItem> {
-  if (env.mode === 'entry') {
-    const dir = await resolveScrapbookDir(ctx, env);
-    return createScrapbookMarkdownAtDir(dir, filename, bodyText, {
-      secret: env.secret,
-    });
-  }
-  return createScrapbookMarkdown(
-    ctx.projectRoot,
-    ctx.config,
-    env.site,
-    env.slug,
-    filename,
-    bodyText,
-    { secret: env.secret },
-  );
+  const dir = await resolveScrapbookDir(ctx, env);
+  return createScrapbookMarkdownAtDir(dir, filename, bodyText, {
+    secret: env.secret,
+  });
 }
 
 export async function uploadDispatch(
@@ -212,19 +114,8 @@ export async function uploadDispatch(
   filename: string,
   buf: Buffer,
 ): Promise<ScrapbookItem> {
-  if (env.mode === 'entry') {
-    const dir = await resolveScrapbookDir(ctx, env);
-    return writeScrapbookUploadAtDir(dir, filename, buf, {
-      secret: env.secret,
-    });
-  }
-  return writeScrapbookUpload(
-    ctx.projectRoot,
-    ctx.config,
-    env.site,
-    env.slug,
-    filename,
-    buf,
-    { secret: env.secret },
-  );
+  const dir = await resolveScrapbookDir(ctx, env);
+  return writeScrapbookUploadAtDir(dir, filename, buf, {
+    secret: env.secret,
+  });
 }
